chore: get cli flags and read config file
Add the initial logic/code for getting the command line flags passed to the
executable, as well as reading the configuration file.

The configuration file contains (for now) the Redis and Database details.

It's also included the file `config.example.json`, which is an example
of the structure of the configuration file.