chore: use package `slog` for structured logging
This commit adds a structured logger (too early?) to the application.

Rationale:

The package `golang.org/x/exp/slog` provides structured logging
functionalities, this comes in handy because having structured logs
is better (for this type of application) when debugging/understanding
what the application is doing, also, the logs are stored to a file on
the file system, which means they persist for later reading/analyzing.