chore: add and modify types at `types.go`
The type for the fields `RPort` and `DbPort` now is `uint18`,
which seems more logical since a port number is an unsigned
16 bits integers.

Add field `Addr` to struct `Flags`, this flag will be used to specify
the address where the HTTP server will be listening.

Add a new struct type `Global`, the purpose of this type is to
work as a glue/namespacing type, for example, it has a field called
`Logger`, this logger then can be used by all attached methods of this
struct type (for example, used by middlewares and handlers.).

BREAKING CHANGE: all places where the modified types are used should update/adapt.