refactor: remove over-contextualization of parameter names for function `NewRedisStorage`
The function `NewRedisStorage` already gives context that the parameters like, for example:
`rPassword` and `rHost` are connection details for the Redis server, and prefixing each parameter
with an **r** is not necessary.

This commit fixes/refactors (I'm sure which verb is more appropiate) it by removing
the letter `r` for each parameter name.