refactor: introduce the field `AppData` in the struct `Config`
The field `AppData` is a string that represents the path of the directory
where data generated by the application will be stored. this way "all" data
will be in a single place.

but what kind of data will be stored there?

Logs data, which brings us to a breaking change (kinda), the field
`Logs` is removed, the application will store logs at `<AppData>/logs`.

NOTE: things like data generated by the database and Redis server won't
be stored here.

NOTE: The configuration example file is updated to reflect the new
configuration structure.