chore: add function for creating appdata dirs
A new function called `CreateAppDataDirs` is introduced,
this function takes in the path from `Config.AppData` and
creates the directories where data generated by the application
will be stored, for now those are logs and images.

Also, this function takes into account whether or not the directories
already exist, if they do, then it won't create them, so previously
generated data won't be lost.