chore: add new column to table `User`
A new column called `Original_Profile_Picture_Id` is being
added to the table `User`, the difference between this new column
and the existing column `Profile_Picture_Id` is: the new column
represents the id of the image before things like resizing, cropping, etc
are applied to it.

The reason why I decided to keep the original image is rather
personal, I think it would be an intesting thing to do.