feat: create package for managing profile images
A new package for managing user profile images is introduced, it
stores the images in the dir path passed to the function `New`, which
is a function from the same package that creates an instance of the struct
`Manager`.

The struct `Manager` provides the following methods of interest:

- InitDirs
- New
- Archive
- ServeImage

`InitDirs` creates the directories where the images (of different formats)
will be stored. The directories will be created inside the directory path
passed to the `New` function.

`New` takes in an instance of `Image`, which represents the image to be processed, and
an instance of `Crop` which represents the values used to crop the image. If everything
goes right, this function returns the id (uuid) of the stored image and a nil error,
otherwise if something goes wrong, an empty string is returned and a non-nil error.

`Archive` moves images identified by an id from the `activeDir` directory to the
`archiveDir` directory. Images inside `archiveDir` won't be able to be requested.

`ServeImage` is a function of type `fiber.Handler` that serves images (identified by their ids) from
the `activeDir`.

Note: `activeDir` and `archiveDir` are constant defined inside the package.

Well, this is an overview of what the package provides.