chore: use SQL Server implementation of `models.Database`
For this to happen, a new field `Database` was added to the struct `Global`,
so handlers and middlewares can access it.

Also, a function is added, which is a utility function to create and test
a connection to the database server using the provided connection details.

The connection details for the database come from the configuration file.