chore: add field `Details` to `ErrorMessage` struct
A new field for including details about the error is added
to the struct `ErrorMessage`, the type of the field is define
using generic types, that is, the struct is taking a parameter list.

As a result of this change, the code using this struct is updated in
order to provided a value for the `details` field.

Also as a minor change, new variables are defined:

- UserIdKey
- IsLoggedInKey

This are used for setting and getting values from `fiber.Ctx.Locals`.