refactor!: make column `Profile_Picture_Id` a default value
After some thinking, I decided to make the column nullable,
and have a default value of an empty string.

The reason for this is: when signing up, a user won't provide
a profile picture, only a name, password and birthdate is required.

As a consequent, the `sqlserver` package was modified, as well as
the queries used when creating a user.