fix: stop column `Profile_Picture_Id` from truncating its values
Initially the column `Profile_Picture_Id` from the table `User` was defined
as `CHAR(16)` because I thought a UUID had a length of 16 characters, when in fact,
it has a length of 36 characters, at least that's the length of the UUIDs the package
`uuid` generates, which are UUID version 4. So, due to the column being `CHAR(16)`, the
values inserted were being truncated, and an error message was being returned.

This fix to this was to simply define the column as `CHAR(36)`.