refactor!: give a more appropiate name to functions
The function `ReadJSONBody` is renamed to `ReadBodyFromRequest`, this
is a more appropiate name because the body is being from the `*fiber.Ctx`
passed as argument.

Also, a new function `ReadJSONBody` is introduced, it's almost the same
as `ReadBodyFromRequest`, but this it's not reading the JSON from a request
body, instead it's the JSON is a `[]byte`.

Note: We could arguably say that the verb `Read` is not appropiate because
we're doing more than reading, we're reading + decoding the JSON. Well, for
now I'll leave those names.