fix: prevent the return of a false non-nil error
How did I found the bug?

The appdata structure is the following (for now):

- <AppData>
  - images
  - logs

I deleted the logs directory, and then started the server with
the hope that `CreatAppDataDirs` would create it again (since that's its job).

For surprise, the `logs` directory was not being created, and the function
that was opening the logs file at `<AppData>/logs/<filename>` was returning
an error because, well, the directory `logs` didn't exist.

So, somehow the function `CreatAppDataDirs` was not creating the `logs` dir,
well, actually, indeed, if I would've had deleted the `images` dir, it wouldn't
have had created it neither.

So, this commit fixes this.