refactor: remove UNIQUE constraint from column `Profile_Picture_Id`
Now the column `Profile_Picture_Id` from the table `User` doesn't
have a UNIQUE constraint, which means duplicate values are acceptable.

The reason for this change is that, when creating a first new user, by default,
the column was receiving a NULL value, this was fine, but when creating a second new
user, the column was, once again, receiving a NULL value, and due to the UNIQUE constraint,
but database was returning an error.

So, the UNIQUE constraint is removed, and the sqlserver database impl was updated
to not expect such error from the database. Also, the error code from the package `models`
related to this constraint is removed.