refactor: optimize all images regardless of their size
Previously, images only greater than `sizeLimit` were optimized,
now all images are optimized regardless of their size.

This also implictly fixes a bug, where images less than `sizeLimit` were
not optimized, and thus the buffer `originalImageBuffer` was nil, and when
getting the size of the image, an error was being returned, and also 0 bytes
images were being saved.

Well, this is an obvious mistake, but I didn't notice it at all, that's
because almost all of the images I was using for testing were greater than
`sizeLimit` :')